requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.0
celery>=5.3.0
redis>=5.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
Pillow>=10.0.0
gunicorn>=21.0.0
whitenoise>=6.6.0
requests>=2.31.0
requests>=2.31.0
